problem

+++ c-skeleton/src/assign_vector_to_centro.c
@@ -6,7 +6,7 @@
 
 
 int assign_vector_to_centroids(cluster_t* centroids, cluster_t* clusters, uint32_t K, squared_distance_func_t distance_type) {
-    int unchanged = 1; // Flag to indicate if the assignment remains unchanged
+    uint64_t unchanged = 1; // Flag to indicate if the assignment remains unchanged
 
 
     // Allocate memory for centroids
